Move presets in each AI tab
